# openapi.yaml
openapi: 3.0.3
info:
  title: Amrutam Telemedicine Backend API
  version: 1.0.0
  description: >
    API for user authentication, doctor availability, consultation booking,
    prescriptions, search, and admin analytics.
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/payments:
    post:
      tags: [Payments]
      summary: Record a payment for a consultation (idempotent, stub gateway integration)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [consultationId, amount, idempotencyKey]
              properties:
                consultationId:
                  type: string
                amount:
                  type: number
                idempotencyKey:
                  type: string
      responses:
        '201':
          description: Payment recorded (or existing one returned for a repeated idempotencyKey)
        '400':
          description: Missing required fields
        '401':
          description: Not authenticated
        '403':
          description: Not the patient on this consultation
        '404':
          description: Consultation not found
  /api/payments/{consultationId}:
    get:
      tags: [Payments]
      summary: Get payment status for a consultation
      security:
        - bearerAuth: []
      parameters:
        - name: consultationId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Payment record
        '404':
          description: No payment found for this consultation
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...